Change container to keep running in the background and change entrypoint.sh to snapshots.sh to make it clear its not actually an entrypoint file any longer

+++ images/wpsnapshots/snapshots.sh
@@ -3,7 +3,11 @@ set -e
 
 maybe_run_wpsnapshots() {
     if [ -e /wpsnapshots/.wpsnapshots.json ]; then
-        ln -s /wpsnapshots/.wpsnapshots.json /root/.wpsnapshots.json
+
+        if [ ! -e /root/.wpsnapshots.json ]; then
+            ln -s /wpsnapshots/.wpsnapshots.json /root/.wpsnapshots.json
+        fi
+
         wpsnapshots "$@"
      else
         echo 'WP Snapshots is not configured, you must run docker-compose wpsnapshots configure <repository>';
